performance/api/requirements.txt
httpx>=0.27.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
